move reqFromSocket to utils so we can require it without the db being init, also cls.binding the socket-message callbacks
